Add WayneD's clone-dest patch from:
https://github.com/WayneD/rsync-patches/blob/master/clone-dest.diff

This patch adds the --clone-dest option that works link --link-dest
but without requiring the metadata of the files to match in order
to be able to share the file's data.

NOTE: this patch is mostly untested because I don't currently have
a btrfs mount to test it out on.  I still need to make sure that a
cloned file gets its destination attributes set correctly after the
clone, for instance.

To use this patch, run these commands for a successful build:

    patch -p1 <patches/clone-dest.diff
    ./configure                         (optional if already run)
    make

based-on: 9a06b2edb0ea1a226bcc642682c07bacd2ea47d3

Changes to be committed:
	modified:   Makefile.in
	modified:   generator.c
	modified:   options.c
	modified:   rsync.1.md
	modified:   rsync.h
	modified:   syscall.c
	modified:   t_stub.c
	modified:   t_unsafe.c